+++ inst/include/colourvalues/colours/colours_hex.hpp
@@ -7,6 +7,8 @@
 #include "colourvalues/alpha/alpha.hpp"
 #include "colourvalues/summary/summary.hpp"
 #include "colourvalues/format/format.hpp"
+#include "colourvalues/colours/generate_colours.hpp"
+#include "colourvalues/output/output.hpp"
 
 namespace colourvalues {
 namespace colours_hex {
@@ -14,94 +16,51 @@ namespace colours_hex {
   // if palette is a string; it's using in-built palettes; nothing to do
   // if palette is function or vectors, force rescaling
 
-  inline Rcpp::StringVector colour_values_to_hex(
-      Rcpp::NumericVector& x,
+  inline SEXP colours_with_summary(
+      Rcpp::NumericVector& full_values,
+      Rcpp::NumericVector& summary,
+      Rcpp::StringVector& summary_values,
       Rcpp::NumericVector& red,
       Rcpp::NumericVector& green,
       Rcpp::NumericVector& blue,
-      Rcpp::NumericVector& alpha,
+      Rcpp::NumericVector& full_alpha,
+      Rcpp::NumericVector& summary_alpha,
       int& alpha_type,
       std::string& na_colour,
       bool& include_alpha
-   ) {
-
-    int n = x.size();
-    double colours = red.size();
-
-    na_colour = na_colour.length() == 9 ? na_colour : na_colour + "FF";
-
-    colourvalues::scale::rescale( x );
-    Rcpp::StringVector hex_strings( n );
-    double step = 1 / ( colours - 1 );  // TODO(test)
-
-    // cublic_b_spoine :: vec.start, vec.end, start.time, step
-    boost::math::cubic_b_spline< double > spline_red(   red.begin(),   red.end(),   0, step );
-    boost::math::cubic_b_spline< double > spline_green( green.begin(), green.end(), 0, step );
-    boost::math::cubic_b_spline< double > spline_blue(  blue.begin(),  blue.end(),  0, step );
-    boost::math::cubic_b_spline< double > spline_alpha( alpha.begin(),  alpha.end(),  0, step );
-
-    double this_x;
-    int i, r, g, b;
-    std::string hex_str;
-
-    for( i = 0; i < n; i++ ) {
-
-      this_x = x[i];
-
-      if ( R_IsNA( this_x) || R_IsNaN( this_x ) ) {
-        hex_strings[i] = na_colour;
-      } else {
-        r = round( spline_red( this_x ) * 255 ) ;
-        g = round( spline_green( this_x ) * 255 );
-        b = round( spline_blue( this_x ) * 255 );
-
-        colourvalues::palette_utils::validate_rgb_spline(r);
-        colourvalues::palette_utils::validate_rgb_spline(g);
-        colourvalues::palette_utils::validate_rgb_spline(b);
-
-        if (include_alpha) {
-          int a;
-          if ( alpha_type == ALPHA_PALETTE ) {
-            a = round( spline_alpha( this_x ) * 255 );
-          } else if (alpha_type == ALPHA_VECTOR ){
-            a = alpha[i];
-          } else {
-            a = alpha[0];  // should be length 5, but all the same
-          }
-          hex_strings[i] = colourvalues::convert::convert_rgb_to_hex(r, g, b, a);
-        } else {
-          hex_strings[i] = colourvalues::convert::convert_rgb_to_hex(r, g, b);
-        }
-      }
-    }
-    return hex_strings;
-  }
-
-  // combining:
-  // - 'format'
-
-  inline Rcpp::List create_summary_output(
-      Rcpp::StringVector& full_hex,
-      Rcpp::StringVector& summary_values,
-      Rcpp::StringVector& summary_hex
   ) {
-    return Rcpp::List::create(
-      _["colours"] = full_hex,
-      _["summary_values"] = summary_values,
-      _["summary_colours"] = summary_hex
+    Rcpp::StringVector full_hex = colourvalues::generate_colours::colour_values_to_hex(
+      full_values, red, green, blue, full_alpha, alpha_type, na_colour, include_alpha
     );
+    Rcpp::StringVector summary_hex = colourvalues::generate_colours::colour_values_to_hex(
+      summary, red, green, blue, summary_alpha, alpha_type, na_colour, include_alpha
+    ); // uses full opacity
+
+    return colourvalues::output::create_summary_output(full_hex, summary_values, summary_hex );
   }
 
-  inline Rcpp::List create_summary_output(
-      Rcpp::StringVector& full_hex,
+  inline SEXP colours_with_summary(
+      Rcpp::NumericVector& full_values,
+      Rcpp::NumericVector& summary,
       SEXP& summary_values,
-      Rcpp::StringVector& summary_hex
-   ) {
-    return Rcpp::List::create(
-      _["colours"] = full_hex,
-      _["summary_values"] = summary_values,
-      _["summary_colours"] = summary_hex
+      Rcpp::NumericVector& red,
+      Rcpp::NumericVector& green,
+      Rcpp::NumericVector& blue,
+      Rcpp::NumericVector& full_alpha,
+      Rcpp::NumericVector& summary_alpha,
+      int& alpha_type,
+      std::string& na_colour,
+      bool& include_alpha
+  ) {
+
+    Rcpp::StringVector full_hex = colourvalues::generate_colours::colour_values_to_hex(
+      full_values, red, green, blue, full_alpha, alpha_type, na_colour, include_alpha
     );
+    Rcpp::StringVector summary_hex = colourvalues::generate_colours::colour_values_to_hex(
+      summary, red, green, blue, summary_alpha, alpha_type, na_colour, include_alpha
+    ); // uses full opacity
+
+    return colourvalues::output::create_summary_output(full_hex, summary_values, summary_hex );
   }
 
   inline SEXP colour_value_hex(
@@ -142,13 +101,13 @@ namespace colours_hex {
       int n_alpha_summary = n_summaries < 5 ? 5 : n_summaries;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::StringVector summary_hex = colour_values_to_hex( summary, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha ); // uses full opacity
-      Rcpp::StringVector full_hex = colour_values_to_hex( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+      return colours_with_summary(
+        x, summary, summary_values, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
 
-      return create_summary_output(full_hex, summary_values, summary_hex );
     }
 
-    return colour_values_to_hex( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+    return colourvalues::generate_colours::colour_values_to_hex( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
   }
 
   inline SEXP colour_value_hex(
@@ -187,12 +146,12 @@ namespace colours_hex {
       int n_alpha_summary = n_summaries < 5 ? 5 : n_summaries;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::StringVector summary_hex = colour_values_to_hex(summary, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha);
-      Rcpp::StringVector full_hex = colour_values_to_hex(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha);
+      return colours_with_summary(
+        x, summary, summary_values, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
 
-      return create_summary_output(full_hex, summary_values, summary_hex );
     }
-    return colour_values_to_hex(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha);
+    return colourvalues::generate_colours::colour_values_to_hex(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha);
   }
 
   inline SEXP colour_value_hex (
@@ -200,7 +159,9 @@ namespace colours_hex {
       Rcpp::NumericMatrix& palette,
       std::string& na_colour,
       bool include_alpha,
-      bool summary = false) {
+      bool summary = false,
+      int is_factor = false
+  ) {
 
     colourvalues::utils::matrix_palette_check( palette );
     int alpha_type = colourvalues::alpha::make_alpha_type( 0, x.size(), palette.ncol() );
@@ -223,13 +184,12 @@ namespace colours_hex {
       int n_alpha_summary = red_size < 5 ? 5 : red_size;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::StringVector summary_hex = colour_values_to_hex( nv, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
-      Rcpp::StringVector full_hex = colour_values_to_hex( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
-
-      return create_summary_output(full_hex, lvls, summary_hex );
+      return colours_with_summary(
+        out_nv, nv, lvls, red, green, blue, alpha, alpha_summary, alpha_type, na_colour, include_alpha
+      );
     }
-    //Rcpp::Rcout << "about to do the conversion - include_alpha: " << include_alpha << std::endl;
-    return colour_values_to_hex( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
+
+    return colourvalues::generate_colours::colour_values_to_hex( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
   }
 
   inline SEXP colour_value_hex(
@@ -238,7 +198,8 @@ namespace colours_hex {
       std::string& na_colour,
       Rcpp::NumericVector& alpha,
       bool include_alpha,
-      bool summary = false) {
+      bool summary = false
+  ) {
 
     int x_size = x.size();
     int alpha_type = colourvalues::alpha::make_alpha_type( alpha.size(), x_size, 0 );
@@ -251,7 +212,7 @@ namespace colours_hex {
 
     colourvalues::palette_utils::resolve_palette( palette, red, green, blue );
 
-    Rcpp::StringVector lvls = Rcpp::sort_unique( x ); // moved outside resolve so can use in a legend
+    Rcpp::StringVector lvls = Rcpp::sort_unique( x ); // moved outside resolve so can use in a summary;
     lvls = na_omit( lvls );
     Rcpp::NumericVector out_nv = colourvalues::utils::resolve_string_vector( x, lvls );
 
@@ -263,13 +224,103 @@ namespace colours_hex {
       int n_alpha_summary = x_size < 5 ? 5 : x_size;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::StringVector summary_hex = colour_values_to_hex( nv, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
-      Rcpp::StringVector full_hex = colour_values_to_hex( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
-
-      return create_summary_output(full_hex, lvls, summary_hex );
+      return colours_with_summary(
+        out_nv, nv, lvls, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
     }
 
-    return colour_values_to_hex( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+    return colourvalues::generate_colours::colour_values_to_hex( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+  }
+
+  /*
+   * i.e, factors
+   */
+  inline SEXP colour_value_hex(
+      Rcpp::IntegerVector& x,
+      Rcpp::StringVector lvls,
+      Rcpp::NumericMatrix& palette,
+      std::string& na_colour,
+      bool include_alpha,
+      bool summary = false
+  ) {
+
+    Rcpp::NumericVector x_nv = Rcpp::as< Rcpp::NumericVector >( x );
+    colourvalues::utils::matrix_palette_check( palette );
+
+    int x_size = x.size();
+    int alpha_type = colourvalues::alpha::make_alpha_type( 0, x_size, palette.ncol() );
+
+    Rcpp::NumericVector red(256);
+    Rcpp::NumericVector green(256);
+    Rcpp::NumericVector blue(256);
+    Rcpp::NumericVector alpha( x.size(), 255.0 );
+
+    colourvalues::palette_utils::resolve_palette( palette, red, green, blue, alpha );
+
+    Rcpp::NumericVector alpha_full = colourvalues::alpha::validate_alpha( alpha, alpha_type, x_size );
+
+    Rcpp::IntegerVector summary_values = Rcpp::sort_unique( x );
+    Rcpp::NumericVector summary_numbers = Rcpp::as< Rcpp::NumericVector >( summary_values );
+
+    if ( summary ) {
+      int x_size = x.size();
+      int n_alpha_summary = x_size < 5 ? 5 : x_size;
+      Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
+
+      return colours_with_summary(
+        x_nv, summary_numbers, lvls, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
+    }
+
+    return colourvalues::generate_colours::colour_values_to_hex( x_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+  }
+
+  /*
+   * i.e, factors
+   */
+  inline SEXP colour_value_hex(
+      Rcpp::IntegerVector& x,
+      Rcpp::StringVector lvls,
+      std::string& palette,
+      std::string& na_colour,
+      Rcpp::NumericVector& alpha,
+      bool include_alpha,
+      bool summary = false
+  ) {
+
+    Rcpp::NumericVector x_nv = Rcpp::as< Rcpp::NumericVector >( x );
+    int x_size = x.size();
+    int alpha_type = colourvalues::alpha::make_alpha_type( alpha.size(), x_size, 0 );
+
+    Rcpp::NumericVector alpha_full = colourvalues::alpha::validate_alpha( alpha, alpha_type, x_size );
+
+    Rcpp::NumericVector red(256);
+    Rcpp::NumericVector green(256);
+    Rcpp::NumericVector blue(256);
+
+    colourvalues::palette_utils::resolve_palette( palette, red, green, blue );
+
+    //Rcpp::Rcout << "x: " << x << std::endl;
+
+    Rcpp::IntegerVector summary_values = Rcpp::sort_unique( x );
+    //Rcpp::Rcout << "summary_values: " << summary_values << std::endl;
+    Rcpp::NumericVector summary_numbers = Rcpp::as< Rcpp::NumericVector >( summary_values );
+
+    if ( summary ) {
+      int x_size = x.size();
+      int n_alpha_summary = x_size < 5 ? 5 : x_size;
+      Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
+
+      // Rcpp::Rcout << "x_nv: " << x_nv << std::endl;
+      // Rcpp::Rcout << "summary_numbers: " << summary_numbers << std::endl;
+      // Rcpp::Rcout << "lvls: " << lvls << std::endl;
+
+      return colours_with_summary(
+        x_nv, summary_numbers, lvls, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
+    }
+
+    return colourvalues::generate_colours::colour_values_to_hex( x_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
   }
 
 } // namespace colours