+++ inst/include/colourvalues/colours/colours_rgb.hpp
@@ -7,77 +7,57 @@
 #include "colourvalues/alpha/alpha.hpp"
 #include "colourvalues/summary/summary.hpp"
 #include "colourvalues/format/format.hpp"
+#include "colourvalues/colours/generate_colours.hpp"
+#include "colourvalues/output/output.hpp"
 
 namespace colourvalues {
 namespace colours_rgb {
 
-// if palette is a string; it's using in-built palettes; nothing to do
-// if palette is function or vectors, force rescaling
-
-  inline Rcpp::NumericMatrix colour_values_to_rgb(
-      Rcpp::NumericVector& x,
+  inline SEXP colours_with_summary(
+      Rcpp::NumericVector& full_values,
+      Rcpp::NumericVector& summary,
+      Rcpp::StringVector& summary_values,
       Rcpp::NumericVector& red,
       Rcpp::NumericVector& green,
       Rcpp::NumericVector& blue,
-      Rcpp::NumericVector& alpha,
+      Rcpp::NumericVector& full_alpha,
+      Rcpp::NumericVector& summary_alpha,
       int& alpha_type,
       std::string& na_colour,
-      bool& include_alpha) {
+      bool& include_alpha
+  ) {
+    Rcpp::NumericMatrix full_rgb = colourvalues::generate_colours::colour_values_to_rgb(
+      full_values, red, green, blue, full_alpha, alpha_type, na_colour, include_alpha
+    );
+    Rcpp::NumericMatrix summary_rgb = colourvalues::generate_colours::colour_values_to_rgb(
+      summary, red, green, blue, summary_alpha, alpha_type, na_colour, include_alpha
+    ); // uses full opacity
 
-    int n = x.size();
-    double colours = red.size();
+    return colourvalues::output::create_summary_output(full_rgb, summary_values, summary_rgb );
+  }
 
-    na_colour = na_colour.length() == 9 ? na_colour : na_colour + "FF";
 
-    colourvalues::scale::rescale( x );
-    int cols = include_alpha ? 4 : 3;
-    Rcpp::NumericMatrix rgb_mat(n, cols);
-    double step = 1 / ( colours - 1 );  // TODO(test)
+  inline SEXP colours_with_summary(
+      Rcpp::NumericVector& full_values,
+      Rcpp::NumericVector& summary,
+      SEXP& summary_values,
+      Rcpp::NumericVector& red,
+      Rcpp::NumericVector& green,
+      Rcpp::NumericVector& blue,
+      Rcpp::NumericVector& full_alpha,
+      Rcpp::NumericVector& summary_alpha,
+      int& alpha_type,
+      std::string& na_colour,
+      bool& include_alpha
+  ) {
+    Rcpp::NumericMatrix full_rgb = colourvalues::generate_colours::colour_values_to_rgb(
+      full_values, red, green, blue, full_alpha, alpha_type, na_colour, include_alpha
+    );
+    Rcpp::NumericMatrix summary_rgb = colourvalues::generate_colours::colour_values_to_rgb(
+      summary, red, green, blue, summary_alpha, alpha_type, na_colour, include_alpha
+    ); // uses full opacity
 
-    // cublic_b_spoine :: vec.start, vec.end, start.time, step
-    boost::math::cubic_b_spline< double > spline_red(   red.begin(),   red.end(),   0, step );
-    boost::math::cubic_b_spline< double > spline_green( green.begin(), green.end(), 0, step );
-    boost::math::cubic_b_spline< double > spline_blue(  blue.begin(),  blue.end(),  0, step );
-    boost::math::cubic_b_spline< double > spline_alpha( alpha.begin(), alpha.end(), 0, step );
-
-    double this_x;
-    int i, r, g, b;
-    std::string hex_str;
-
-    Rcpp::IntegerMatrix na_mat = colourvalues::convert::convert_hex_to_rgb( na_colour );
-
-    for( i = 0; i < n; i++ ) {
-
-      this_x = x[i];
-
-      if ( R_IsNA( this_x) || R_IsNaN( this_x ) ) {
-        rgb_mat(i, _) = na_mat;
-
-      } else {
-        r = round( spline_red( this_x ) * 255 ) ;
-        g = round( spline_green( this_x ) * 255 );
-        b = round( spline_blue( this_x ) * 255 );
-
-        r = colourvalues::palette_utils::validate_rgb_range( r );
-        g = colourvalues::palette_utils::validate_rgb_range( g );
-        b = colourvalues::palette_utils::validate_rgb_range( b );
-
-        if (include_alpha) {
-          int a;
-          if ( alpha_type == ALPHA_PALETTE ) {
-            a = round( spline_alpha( this_x ) * 255 );
-          } else if (alpha_type == ALPHA_VECTOR ){
-            a = alpha[i];
-          } else {
-            a = alpha[0];  // should be length 5, but all the same
-          }
-          rgb_mat(i, _) = Rcpp::NumericVector::create(r,g,b,a);
-        } else {
-          rgb_mat(i, _) = Rcpp::NumericVector::create(r,g,b);
-        }
-      }
-    }
-    return rgb_mat;
+    return colourvalues::output::create_summary_output(full_rgb, summary_values, summary_rgb );
   }
 
   // in this function the colour vectors will already be scaled [0,1]
@@ -119,16 +99,20 @@ namespace colours_rgb {
       int n_alpha_summary = n_summaries < 5 ? 5 : n_summaries;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::NumericMatrix summary_rgb = colour_values_to_rgb(x, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
-      Rcpp::NumericMatrix full_rgb = colour_values_to_rgb( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
-      return Rcpp::List::create(
-        _["colours"] = full_rgb,
-        _["summary_values"] = summary_values,
-        _["summary_colours"] = summary_rgb
+      return colours_with_summary(
+        x, summary, summary_values, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
       );
+
+      // Rcpp::NumericMatrix summary_rgb = colourvalues::generate_colours::colour_values_to_rgb(x, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
+      // Rcpp::NumericMatrix full_rgb = colourvalues::generate_colours::colour_values_to_rgb( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+      // return Rcpp::List::create(
+      //   _["colours"] = full_rgb,
+      //   _["summary_values"] = summary_values,
+      //   _["summary_colours"] = summary_rgb
+      // );
     }
 
-    return colour_values_to_rgb( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+    return colourvalues::generate_colours::colour_values_to_rgb( x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
   }
 
   inline SEXP colour_value_rgb(
@@ -166,16 +150,20 @@ namespace colours_rgb {
       int n_alpha_summary = n_summaries < 5 ? 5 : n_summaries;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::NumericMatrix summary_rgb = colour_values_to_rgb(summary, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
-      Rcpp::NumericMatrix full_rgb = colour_values_to_rgb(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
-      return Rcpp::List::create(
-        _["colours"] = full_rgb,
-        _["summary_values"] = summary_values,
-        _["summary_colours"] = summary_rgb
+      return colours_with_summary(
+        x, summary, summary_values, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
       );
+
+      // Rcpp::NumericMatrix summary_rgb = colourvalues::generate_colours::colour_values_to_rgb(summary, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
+      // Rcpp::NumericMatrix full_rgb = colourvalues::generate_colours::colour_values_to_rgb(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+      // return Rcpp::List::create(
+      //   _["colours"] = full_rgb,
+      //   _["summary_values"] = summary_values,
+      //   _["summary_colours"] = summary_rgb
+      // );
     }
 
-    return colour_values_to_rgb(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+    return colourvalues::generate_colours::colour_values_to_rgb(x, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
   }
 
   inline SEXP colour_value_rgb (
@@ -205,16 +193,20 @@ namespace colours_rgb {
       int n_alpha_summary = red_size < 5 ? 5 : red_size;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::NumericMatrix summary_rgb = colour_values_to_rgb( nv, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
-      Rcpp::NumericMatrix full_rgb = colour_values_to_rgb( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
-      return Rcpp::List::create(
-        _["colours"] = full_rgb,
-        _["summary_values"] = lvls,
-        _["summary_colours"] = summary_rgb
+      return colours_with_summary(
+        out_nv, nv, lvls, red, green, blue, alpha, alpha_summary, alpha_type, na_colour, include_alpha
       );
+
+      // Rcpp::NumericMatrix summary_rgb = colourvalues::generate_colours::colour_values_to_rgb( nv, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
+      // Rcpp::NumericMatrix full_rgb = colourvalues::generate_colours::colour_values_to_rgb( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
+      // return Rcpp::List::create(
+      //   _["colours"] = full_rgb,
+      //   _["summary_values"] = lvls,
+      //   _["summary_colours"] = summary_rgb
+      // );
     }
 
-    return colour_values_to_rgb( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
+    return colourvalues::generate_colours::colour_values_to_rgb( out_nv, red, green, blue, alpha, alpha_type, na_colour, include_alpha );
   }
 
   inline SEXP colour_value_rgb(
@@ -246,16 +238,104 @@ namespace colours_rgb {
       int n_alpha_summary = x_size < 5 ? 5 : x_size;
       Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
 
-      Rcpp::NumericMatrix summary_rgb = colour_values_to_rgb( nv, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
-      Rcpp::NumericMatrix full_rgb = colour_values_to_rgb( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
-      return Rcpp::List::create(
-        _["colours"] = full_rgb,
-        _["summary_values"] = lvls,
-        _["summary_colours"] = summary_rgb
+      return colours_with_summary(
+        out_nv, nv, lvls, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
+
+      // Rcpp::NumericMatrix summary_rgb = colourvalues::generate_colours::colour_values_to_rgb( nv, red, green, blue, alpha_summary, alpha_type, na_colour, include_alpha );
+      // Rcpp::NumericMatrix full_rgb = colourvalues::generate_colours::colour_values_to_rgb( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+      // return Rcpp::List::create(
+      //   _["colours"] = full_rgb,
+      //   _["summary_values"] = lvls,
+      //   _["summary_colours"] = summary_rgb
+      // );
+    }
+
+    return colourvalues::generate_colours::colour_values_to_rgb( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+  }
+
+  /*
+   * i.e, factors
+   */
+  inline SEXP colour_value_rgb(
+      Rcpp::IntegerVector& x,
+      Rcpp::StringVector lvls,
+      Rcpp::NumericMatrix& palette,
+      std::string& na_colour,
+      bool include_alpha,
+      bool summary = false
+  ) {
+
+    Rcpp::NumericVector x_nv = Rcpp::as< Rcpp::NumericVector >( x );
+    colourvalues::utils::matrix_palette_check( palette );
+
+    int x_size = x.size();
+    int alpha_type = colourvalues::alpha::make_alpha_type( 0, x_size, palette.ncol() );
+
+    Rcpp::NumericVector red(256);
+    Rcpp::NumericVector green(256);
+    Rcpp::NumericVector blue(256);
+    Rcpp::NumericVector alpha( x.size(), 255.0 );
+
+    colourvalues::palette_utils::resolve_palette( palette, red, green, blue, alpha );
+
+    Rcpp::NumericVector alpha_full = colourvalues::alpha::validate_alpha( alpha, alpha_type, x_size );
+
+    Rcpp::IntegerVector summary_values = Rcpp::sort_unique( x );
+    Rcpp::NumericVector summary_numbers = Rcpp::as< Rcpp::NumericVector >( summary_values );
+
+    if ( summary ) {
+      int x_size = x.size();
+      int n_alpha_summary = x_size < 5 ? 5 : x_size;
+      Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
+
+      return colours_with_summary(
+        x_nv, summary_numbers, lvls, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
       );
     }
 
-    return colour_values_to_rgb( out_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+    return colourvalues::generate_colours::colour_values_to_rgb( x_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
+  }
+
+  /*
+   * i.e, factors
+   */
+  inline SEXP colour_value_rgb(
+      Rcpp::IntegerVector& x,
+      Rcpp::StringVector lvls,
+      std::string& palette,
+      std::string& na_colour,
+      Rcpp::NumericVector& alpha,
+      bool include_alpha,
+      bool summary = false
+  ) {
+
+    Rcpp::NumericVector x_nv = Rcpp::as< Rcpp::NumericVector >( x );
+    int x_size = x.size();
+    int alpha_type = colourvalues::alpha::make_alpha_type( alpha.size(), x_size, 0 );
+
+    Rcpp::NumericVector alpha_full = colourvalues::alpha::validate_alpha( alpha, alpha_type, x_size );
+
+    Rcpp::NumericVector red(256);
+    Rcpp::NumericVector green(256);
+    Rcpp::NumericVector blue(256);
+
+    colourvalues::palette_utils::resolve_palette( palette, red, green, blue );
+
+    Rcpp::IntegerVector summary_values = Rcpp::sort_unique( x );
+    Rcpp::NumericVector summary_numbers = Rcpp::as< Rcpp::NumericVector >( summary_values );
+
+    if ( summary ) {
+      int x_size = x.size();
+      int n_alpha_summary = x_size < 5 ? 5 : x_size;
+      Rcpp::NumericVector alpha_summary( n_alpha_summary, 255.0 );
+
+      return colours_with_summary(
+        x_nv, summary_numbers, lvls, red, green, blue, alpha_full, alpha_summary, alpha_type, na_colour, include_alpha
+      );
+    }
+
+    return colourvalues::generate_colours::colour_values_to_rgb( x_nv, red, green, blue, alpha_full, alpha_type, na_colour, include_alpha );
   }
 
 } // namespace colours